Cleanup all forge usages of ClientHooks and mark for removal.
